+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Exposed `get_blocks` function on the v2 client which returns a stream of blocks `ArrivedBlockInfo`. This endpoint can be used to listen for new blocks arriving on chain.
+
 ## 9.0.1
 
 - Fixes a bug where using `serde` to convert `BlockItemSummary` would result in data loss with regards to transaction sponsor details.